Back-End.C#/Aulas/AulaTestes.NET/AulaTestes.NET/Account.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace AulaTestes.NET
{
    public class Account
    {
        public double m_balance {  get; set; }
        public void Withdraw(double amount)
        {
            if (m_balance >= amount)
            {
                m_balance -= amount;
            }
            else
            {
                throw new ArgumentException(nameof(amount), "Saldo insuficiente");
            }
        }
    }
}
